backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
websockets>=11.0.0
aiohttp>=3.8.0
python-dotenv>=1.0.0